+++ src/settings.ts
@@ -126,6 +126,9 @@ class ProblemSection extends SettingsSection {
     get templateFile(): string {
         return this.get('templateFile') as string;
     }
+    get problemFilePath(): string {
+        return this.get('problemFilePath') as string;
+    }
 }
 
 class SidebarSection extends SettingsSection {