+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the "cph-ng" extension will be documented in this file.
 
 ## [Unreleased]
 
+- **feat**: Add problem file path configuration
+
 ## 0.0.3
 
 - **feat**: Add support for saving unsaved changes in the editor before